backend/requirements.txt
# Web framework
fastapi==0.135.1
uvicorn==0.41.0

# Database
sqlalchemy==2.0.48
alembic==1.18.4

# Async HTTP
aiohttp==3.13.3
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx==0.28.1

# Task queue
celery[redis]==5.6.2
redis==6.4.0
flower==2.0.1

# Analytics
pandas==3.0.1
duckdb==1.4.4
numpy==2.4.2

# LLM
groq==1.0.0

# Config
python-dotenv==1.2.2
pyyaml==6.0.3

# Utilities
python-dateutil==2.9.0.post0
tenacity==9.1.4
